move nodestate
