+++ src/griptape_nodes/retained_mode/managers/directed_graph.py
@@ -2,7 +2,6 @@ from __future__ import annotations
 
 import graphlib
 import logging
-from enum import StrEnum
 
 logger = logging.getLogger("griptape_nodes")
 
@@ -59,14 +58,3 @@ class DirectedGraph:
     def get_topological_sorter(self) -> graphlib.TopologicalSorter[str]:
         """Create a TopologicalSorter from the current graph structure."""
         return graphlib.TopologicalSorter(self._predecessors)
-
-
-class NodeState(StrEnum):
-    """Individual node execution states."""
-
-    QUEUED = "queued"
-    PROCESSING = "processing"
-    DONE = "done"
-    CANCELED = "canceled"
-    ERRORED = "errored"
-    WAITING = "waiting"
